Extract some modules out of file.rs
