+++ crates/compiler/test_derive/src/util.rs
@@ -22,7 +22,8 @@ use roc_constrain::expr::constrain_decls;
 use roc_debug_flags::dbg_do;
 use roc_derive::DerivedModule;
 use roc_derive_key::{DeriveBuiltin, DeriveError, DeriveKey, Derived};
-use roc_load_internal::file::{add_imports, LoadedModule, Threading};
+use roc_load_internal::file::{add_imports, Threading};
+use roc_load_internal::module::LoadedModule;
 use roc_module::symbol::{IdentIds, Interns, ModuleId, Symbol};
 use roc_region::all::LineInfo;
 use roc_reporting::report::{type_problem, RocDocAllocator};
